+++ utils/middleware/jwt.go
@@ -9,7 +9,7 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-func GenerateToken(id int, role string) (s string, err error) {
+func GenerateToken(id int64, role string) (s string, err error) {
 	token := jwt.New(jwt.SigningMethodHS256)
 	claims := token.Claims.(jwt.MapClaims)
 	claims["id"] = id
@@ -25,10 +25,10 @@ func GenerateToken(id int, role string) (s string, err error) {
 	return
 }
 
-func ExtractToken(c *fiber.Ctx) (id int, role string) {
+func ExtractToken(c *fiber.Ctx) (id int64, role string) {
 	user := c.Locals("user").(*jwt.Token)
 	claims := user.Claims.(jwt.MapClaims)
-	id = int(claims["id"].(float64))
+	id = int64(claims["id"].(float64))
 	role = claims["role"].(string)
 
 	return
